mwana/requirements/libs.txt
# django>=1.2.3
Django-1.2.3.tar.gz
# nose>=0.11.2
nose-0.11.2.tar.gz
# django-nose>=0.1.2
django-nose-0.1.2.tar.gz
# djtables>=0.1.1
djtables-0.1.1.tar.gz
# djappsettings>=0.1.0
djappsettings-0.1.0.tar.gz
# RapidSMS:
rapidsms-gefd0256.tar.gz
# pytz:
pytz-2010l.tar.gz
# pygsm must be installed separately because it requires that pytz be
# installed first
# django-uni-form (for xforms support)
django-uni-form-0.7.0.tar.gz
# django-eav (for xforms support)
django-eav-e8b5849.tar.gz
# rapidsms-xforms
rapidsms-xforms.tar.gz
# pyserial
pyserial-2.5.tar.gz
# pygsm must be installed separately because it requires that pytz and pyserial
# be installed first
# pygrowup (from https://github.com/ewheeler/pygrowup)
pygrowup-b10ba44.tar.gz
